+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [2.13.2] - 2021-02-10
+
 ### Changed
 
   - Fulltext search enhanced